+++ src/tools/sendchangelog.ts
@@ -22,7 +22,9 @@ if (process.env["NODE_ENV"] !== "test") {
 
   // send email
   const toAddresses: string[] = ["linxiaoyi@gagogroup.com", "jiangwei@gagogroup.com", "tangyongtao@gagogroup.com",
-    "huangtaihu@gagogroup.com", "heming@gagogroup.com", "cuixiangchang@gagogroup.com", "jialongfei@gagogroup.com"];
+    "huangtaihu@gagogroup.com", "libiyang@gagogroup.com", "cuixiangchang@gagogroup.com", "jialongfei@gagogroup.com",
+    "gaoqiang@gagogroup.com", "wuenping@gagogroup.com", "wanghui@gagogroup.com", "qiaolei@gagogroup.com",
+    "yangdonglai@gagogroup.com"];
 
   const packageJson: any = require(path.resolve("./package.json"));
   const currentVersion: string = String(packageJson.version);